requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pyswisseph==2.10.3.2
gunicorn==21.2.0